fix(map): minor cursor and folds
when inserting pairs (`()`, `""`, etc.) you can now move past the
autoinserted second half of the pair with `<c-l>` in insert mode,
regardless of the kind of pair.

`fold less` is now a mnemonic for reducing folds with `zl`, which I find
better than `fold reduce` (it's not even in the right order in english).