feat(plug): proper file paths in lualine
It took me a little to figure out how to customize the components for
lualine, but now that I figured it out I do have to say that it is
pretty straight forward and easy to play with.  I can't believe I
haven't fixed this a long time ago, since it has been bothering me for
quite a while.