ref(plugin): declare all commands in lua/commands.lua
In the past I was using the `config` field in packer to define some
commands related to the plugin.  This tends to be kind of annoying when
I want to redefine those commands, as they are no longer all in one
place.  It theoretically saves me some cleanup when I remove a plugin,
but I guess I just have to come up with another way of ensuring that.
In the end I don't think that is all that important - it has not been a
problem so far at all.