new(plug): reticle, cursor-line follows me
This plugin automatically moves the cursor-line with the active window,
but also allows me to leave the number-line active on the other windows.
So I can see where the cursor is on the other windows without it
distracting me from the active window.  Since I have sometimes wondered
why the cursor wasn't moving simply because I was looking at an inactive
window.