new(plug): fugitive
While I tried to stay away from vimscript plugins, this one has to be an
exception.  For one it is better maintained than many lua plugins, so it
very much goes against the overall pattern with that.

On top of that, people tend not to want to reimplement parts of it that
they feel is unnecessary, especially given this plugin's high quality.

The main reason I wanted it now was to commit from within nvim.  Now I
can stage hunks with gitsigns, stage whole files with telescope and I
can stage everything with `:G add -A`.  All this I can commit and write
the commit message, however long or short, right inside my editor.

So on the tmux tabs that I have open I mostly don't have to split
another shell anymore now.  And for the few times that I do I can run
:TERM

This might work.