new(plug): firenvim from browser
This allows me to use nvim from within my browser.  It doesn't fully
work on Vivaldi - (<eyeroll>) with Jira/Confluence just not working -
plus many input fields are just too small for this to be useful
(e.g. reddit) - statusline etc takes up too much space.  But it is a
cool idea and it is nice to be able to play with it when I am in the
mood.

https://chrome.google.com/webstore/detail/firenvim/egpjdkipkomnmjhjmdamaniclmdlobbo