feat(readme): update all
I have been keeping the old description around for way too long.  I have
been firmly in the Linux and neovim camp so all of that is just a waste
of Readme real estate.

I also switched the branching system here, which is what prompted me to
update the Readme.