feat(bash): updateAll
I tried this before as a function, but for some reason this wasn't
working.  As in, the function wasn't available after being defined.  I
have no idea why that didn't work, but with an alias it works.  So now,
while this is less user friendly, I'll be using this.

I also added rustup to the mix.  Since I have been forgetting about it.