new(pacman,bash): cargo-update
Cargo is not a system package manager, but there are some tools that I
install using cargo - from the official instructions so there is nothing
wrong with doing that.  It works a bit like pip in that way.

What is missing is a way to update all of these tools at once, and this
is that.  It works a lot better than pip in that way.