feat(config): k9s better pretty json
Mark the start of the stream visually, since the stream doesn't just
show you the new things in tmux, which can make it hard to see what is
part of the currently running container and what is left over from the
last one that was running the older version.

Also colorize the pretty.