feat(map): find files from window commands
Sometimes I just find myself pressing <C-W> a lot and using the mappings
I have set up with it.  Usually this is about arranging my buffers in
some way that I want them to be so I will be able to focus in the next
step.  Finding files is part of that workflow, so I might as well make
it part of that family of mappings.

I may make an effort to use the hydra more for this stuff, but for some
reason it isn't really working for me on this front.  Either way, if it
would work I would add it there as well.