new(plug): treesitter autopair for html
Autopair in html has been leaving something to be desired so far.  This
plugin seems to do the trick - and it's from the same author as my other
autopair plugin, so I hope they will play nicely together.

Required a bit of rerunning TSUpdate, but it is not working so far.