chore(init): up to date way to do mappings
I am not sure when this convention switched, but now I see in plugin
definitions that `vim.keymap.set` is used to create keymaps, where
previously this was via a `vim.api` route.

Backwards compatibility seemed to have been preserved, but since this
seems to be the common way of doing things now I figure it might be
preferable to do it this way.