new(plug): more TS based navigation - tree-climber
This will take a little getting used to, and not sure that it will be
very useful, especially with the existing TS based navigations I already
have installed.  However, might be nice to have.  One use case that I
see is when I want to traverse the branches of a conditional, since
each branch is a TS child of that if node.  Not sure why, but that
particular case has been eluding my optimisation attempts.