feat(setup): neovim and rustup install
Rustup is really part of pacman, since it doesn't require any
configuration.

Neovim install will require me to close the application in the middle,
but other than that it can be added to my setup scripts no problem.  It
really should be in there I think, but we'll see how that goes.

The README.md reflects this change in setup.