feat(hydra): clean up unused
I can't seem to make use of hydra as much as I thought I would.  Perhaps
I just need to walk before I run with it, or perhaps it's not for me.

The issue that makes me remove it is that it does take up keymaps, which
sometimes interferes.  All the recent times I have edited the hydras has
been about these conflicts, but I have not used a hydra in a long time.
So it is time to remove those.

The only exception may be the window management hydra, which has not
been interfering with anything so far.  I think I even did use it a
little while ago to adjust some splits.  I don't use it often and the
more complex parts of it I don't remember, but it still has a positive
impact in the bottom line.

All other hydras, however, have been more annoying than anything else
over the past few months.

This is kind of sad, because I think they have a lot of potential.  They
sound like such a good idea.  They just don't work for me right now, at
all.