feat(plug,map): proper ; and , with TS support
Treesitter objects allow you to repeat movements with ';' and ','.  This
is working great so far and I am hoping to extend it with any other
movements I might have in other plugins, since it apparently also allows
that.

This is also when I learned that ',' is the inverse of ';'.  Despite
intentionally looking for an inverse of ';' I did not find out about
this until today.  Which required me to fix some conflicts in my
mappings for this to work properly.  It also prompted me to remove most
hydras in the last commit (4644211).