feat(bash): clean and update docker/kubernetes
There were a bunch of old commands for minikube and docker-compose left
over from back when.  I don't use those anymore.

But I do sometimes want to have a json prettyfied log for a deployment
or job.  And I don't always want to have to do this inside k9s.