feat(map,commands): better git and telescope undo
For git undoing a hunk and resetting it are kind of similar, so they can
be on the same key.  The shift modifier also makes it less likely that I
would reset a hunk by accident, since that is a destructive action.

And the now free key can be used for telescope undo.