fix(config/tmux): real colors, finally
It took me a while, because a lot of the recommended fixes didn't work
for me.  Or maybe it was that I have to kill all tmux sessions before
the color change can take effect and instead of doing that I was going
through a million recommendations what how to fix it.

Either way, it bloody works now!  Awesome!