feat(tmux): plugin manager
I thought I had started using this plugin manager in the past, but
evidently not.  So here it is.

Automatically install the manager if it is not already installed, but
still have to install the plugins.  Minor inconvenience that I will fix
some other time.

For now let's start playing with this.