feat(tmux): yank plugin
This works decent right now.  I am not sure if I want to stick with this
style of selecting, but it definitely is better than the old way *in the
tmux select mode*.  However, it does interfere with how I used to copy
using the mouse.  If I don't find a solution for that (there should be,
I think), I might switch back, and also disable the mouse again.