feat(bash/util): tre ignores symlinks
We use symlinks at work to have the local subdirectory better reflect
the file-tree of the final docker image that will be created from it.
However, usually when I to use tree these are irrelevant, so I would
prefer to skip them.

There is no direct way to cut symlinks from tree, but in the default
format they have a -> in them.