new(plug): indent guides
This is something that is somewhat useful in python, and kind of nice in
most languages.

However, I never really liked the defaults.  They felt visually invasive
to me.  This highlight based stuff is a lot easier on the eyes, though
it is colorscheme specific.  Good thing I don't switch colorschemes
pretty much ever.

The only issue so far is that it overwrites the highlight of the
cursorline.  Maybe I can figure something out on that front - or I will
just deactivate this again.  Or switch back to the defaults once I get
used to it.  We'll see.