feat(plug): firenvim configure
I decided to look at this again, and so long as it does not
automatically appear on a page I can use it.  This is what takeover
does.  Now I just manually start it only when I want it.

I also think the firenvim command line works a little better.

Other settings are just defaults.

I still struggle with using it on Jira/Confluence pages, but it works in
Slack and GitHub and I am still looking for more places where it
does/doesn't work.