fix(setup): no sudo on yay
This used to get you a warning, but now it seems to actually fail.  The
makepkg command will fail if you run it as sudo because it is possible
to irreversibly damage the system that way.

I should have fixed this a long time ago, really, but now they made me.