fix(setup/config): neovim firenvim, plus doc
Despite the fact this command should run with PackerSync/startup, this
was not running properly when I installed on a new machine.  There is no
problem to just run it on install anyway though.

Also, it is nice to have a link to explain why I am running those
commands on install.  I remember the lsp link I put in now usually took
me a while to find in the past, but I always thought "now I understand
this and next time I won't even need to look for it, or I'll remember
anyway".