feat(config/starship): exit status on failure
Display exit status iff the process did not exit with 0.  Additional
information can sometimes be useful.  Especially when writing your own
scripts where this can be a very quick and convenient way to fail your
script.

Of course, any tool worth using will usually exit with stderr output,
but a quick script that I throw together might now.

This does not interfere with anything else, so it is a free addition to
the info I get on failure.