feat(bash|setup): no more ack, repgrep instead
So, I just realised I have not been using ack in a long time.  I use
ripgrep and fd, which has been working very well for me.  I don't seem
to have used any of the scripts I put together with ack.  Partly,
probably, because of config files that come with fd and ripgrep.  Not
that I couldn't have done that with ack and find, but I like Rust tools.

Either way, I may be going to use repgrep in the future for some things,
if I remember that it exists.