feat(config/kde): kwin overview less annoying
Sometimes when switching to Desktop 1 I keep calling the KWin Overview
function.  Sometimes it keeps happening, sometimes it doesn't happen at
all.  When it does happen (especially repeatedly) it is very annoying.

I don't actually use this, but I decided to just remap it rather than
removing it from my shortcuts entirely.  It is now similar to the zoom
in neovim - which, come to think of it, I haven't used in a while either.
Maybe if I use tiling a little bit more I will start having more use for
that, and this may be useful in itself..