feat(map): better reverse git jump
When jumping to the closest git hunk before the cursor I prefer not
having to also press and let go of shift.

This does require using `GG` to get to the bottom of the file.  Which
considering how basic it is takes some getting used to, but I don't see
why not.  It is not bad muscle memory (in vanilla it doesn't break
anything).  I've actually been wondering why the different number of
keys between `gg` and `G` - good that I fix that inconsistency now :)