feat(ftplugin/gitcommit): more undo breaks
Opening and closing brackets, quotations, etc, are often very similar to
half-sentences.  So they should trigger an undo break at the same time.

The only issue here is auto-pair.  This prevents me from properly
remapping the opening part, however, since I have my <c-l> in insert
mode I can add the undo break to that.  Of course that means any changes
to that remap have to be duplicated for ftplugin/gitcommit, and I won't
remember that and notice way later, but I think that is ok.  It is a
small quality of life improvement.