feat(bash/git): cleanup and color
I usually also want the short hash, but with colors that is easy to fit
into the message without creating information overload.

As a result I only need the default git log format and an extended
version that builds on the default, just for `-p`.

And I just realised that I have created some neat coloring for whole
commit messages a long time ago, not sure why I never use it...  I'll
keep it a little longer, maybe this time I will remember that it exists.

Eventually I may use git alias for this rather than bashrc, but then
again, why?  It works great like this so far.