clean(plug/obsession): I have never used sessions
I tried them a bunch of times, but I never could make them convenient.

Perhaps it is a memory issue.  When opening a session the first thing I
think is 'wtF was I working on here?' and there was no good way to write
this down - especially not in conjunction with the notes I was already
writing down.