feat(kde): jump to obsidian shortcut
These are two shortcut changes.  I was hoping to combine them, but
didn't figure out how.

* start obsidian, on virtual desktop 7 (since that was unused), with
  meta+shift+d
* jump to obsidian (on its virtual desktop) from wherever you are with
  meta+d

This way I can leave the program on _any_ virtual desktop and not even
care which.  I think I will experiment with this.

I'd like a lot more control, though.  This probably won't scale the way
it is, but the alternatives that I am aware of duplicate windows, which
is not a solution either.