wip(nvim): start splitting up lazy plugins
I should be splitting up my plugins into files.

This way I might be reworking some of them.

Here I start with the color scheme and set up a way to toggle between
the - still functional - old setup and the - not yet functional - new
setup.  This way I can commit WIP without breaking anything.

Turns out there is also a bunch of formatting that I committed with
this.  But, such is life, for now.