feat(config/git): transfer some aliases from git.bash to gitconfig
Some of these depend on some bash logic that I have to replicate in any
shell I use if I want to switch - e.g. to fish.  But keeping it in the
gitconfig as much as possible should simplify some things if I end up
using more than one shell for a little while, or want to make future
transitions easier.