feat(config/git): add the last aliases from b ash
One of these is not quite right, which uses the githubcli to get the
base branch for any PR that I might be on right now.  I will have to
figure that out later, but for now that is fine with `development` hard
coded.

The issue with that is that I would prefer not to make assumptions about
the underlying shell inside my gitconfig where possible, and even less
so about the functions available.