feat(config/git): prune well i.e. remove local copy of branch too
For some reason people tend to suggest a ton of different ways to remove
local branches that are left over from remote branches.  See,
e.g. https://stackoverflow.com/questions/13064613/how-to-prune-local-tracking-branches-that-do-not-exist-on-remote-anymore

Some of these answers delete local branches not on remote, and tend to
be a bunch longer.  But the answer I chose to copy from is short, only
deletes local branches that used to have a remote branch but don't
anymore, and otherwise is exactly what I need.

I can first list the branches, inspect them individually if I want,
store the list in a file or something if I want to keep any branches,
etc.  This is great.

Did I mention it does not delete local-only branches?  So stuff I was
not ready to push yet stays.