feat(config/git): find next commit in pager
When viewing logs with text and diff I am not aware of a good way in my
pager to jump to the next commit.  You can use `n` and `N` to jump to
the next and previous hunk if you haven't searched for anything yet, but
nothing for commits.

Now I can to search for `=-=-=` and jump to the next commit.