fixup! fix(config/alacritty): remove legacy
turns out, it really was the difference in the way the different
computers display on the same monitor.  I may address this in the future
with some direct arguments `--option font.size=12.0` on the laptop, but
we'll see.  Perhaps there is a better way, perhaps there is something in
the kde I can adapt, we'll see.