feat(config/fish): close multiple github prs
Specific to work, when there are multiple PRs that are all ready to
merge, but after merging any one you have to update the remaining and
wait for tests to finish again, it is quite tedious.

This is very convenient when you have another copy of the repo on your
local and it can run there, while you are doing real work on your main
copy.